Use BTreeMap instead of HashMap for config.yaml's hash (#2991)
`iotedged` detects changes to `config.yaml` by creating an hash of it's in-memory representation, saving it to disk, and comparing it against the latest next time it starts up. The problem is that several structs which make up our in-memory representation of `config.yaml` use `HashMap` as a dictionary. `HashMap` cannot guarantee that different maps with the same keys will serialize in the same order, so two semantically equivalent `HashMap`s might not compare equal. The impact is that `iotedged` might think that `config.yaml` changed (and remove/restart all modules as a result) when it didn't.

This change moves the offending types to `BTreeMap`, which has the guarantee we need. It also removes the custom `serialize_ordered` attribute on `HashMap`, which just copies the `HashMap` to a `BTreeMap` anyway so using `BTreeMap` to begin with is simpler.