Fixing arm32 new e2e tests (#2956)
Previously arm32 new e2e tests assume that the repo is already there. Somehow our pi got into a bad state, and the git command failed, failing the whole pipeline. 

This change changes the command to do the same thing that we wanted to do (delete any file that was owned by root in the current folder) but without making any assumptions.