Added session config and message limits (#3063)
**Context:**
In order to help with memory consumption for offline or slow subscribers, we need to introduce limits on message queue size and inflight message queue size.

The main complexity here is that queues live in `SessionState`, so we better put `SessionConfig` there as well. But `SessionState` also participates in persistence, but we don't want to persist `SessionConfig`, since it can change b/w broker restarts.

**Solutions:**
There were two ways to work this out:
1. Add an extra layer (data structure) `SessionSnapshot` that would not have `SessionConfig` and be used for persistence. In `BrokerBuilder`, on broker load, convert `SessionSnapshot` into "runtime" `SessionState` by adding `SessionConfig` to it. This is more complex approach.
2. Load `SessionState` from persisted state with default config values, but then ensure to set actual `SessionConfig` in the `BrokerBuilder`. This is less complex but less "rusty" approach.

In this PR I'm attempting approach number 1.