Add watchdog to checkin and consolidate replicated files (#3200)
Since we added the watchdog we need to enable the build, format, and clippy in the checkin gate. I have discussed with Vadim and we think this is a good opportunity to consolidate build scripts for mqtt / edgelet that have nearly identical logic. Thus, I have consolidated the rust install, build, and format scripts under the path /scripts/linux/generic-rust/[name].sh.

The watchdog checkin is not enabled right now because the yaml needs to exist in the iiot branch first. Once it is merged I will create the checkin pipeline, verify it works, then add it to the gate.