[PnP] Add ModelIdStore for EdgeHub to store modelId per device identity (#3216)
This PR adds the ModelIdStore for EdgeHub. This store will store modelIds per device identity.

Every time an identity (module or device) connects to edgeHub with a modelId, we will store the modelId in this store.
Currently, we only implement this change for MQTT.

We will then get the modelId when connecting to upstream and use it to create the device or module client with the modelId if it exists. This will register the identity as PnP in the PnP service.