make the broker send the current subscription list when a client connects (#3239)
The broker and edgehub needs a common state of the subscriptions of a client. in case of persistent sessions, when a client comes back, it already has a (previously established) subscription state. To avoid state-mismatches, it is best to store this state in one place, and the best place for that is in the broker. That means that when a client comes back, EdgeHub will not know the already established subscriptions. In order to let it know, the broker should send it when a client connects.