[k8s public preview] Pull in 1.0.9.4 fixes, update images and Helm charts (#3272)
* Fix edge agent connection 1.0.9 (#3172)

* Recreate connection after every exception

* Fix: Edgelet unable to pull using certain passwords - cherry-pick into 1.0.9. (#3206) (#3209)

Some passwords (especially generated passwords) can be used by docker CLI, but not by Edgelet. There is a chance of this  happening when passwords contain `?` or `~`.

It's not explicitly stated in the [Docker API](https://docs.docker.com/engine/api/v1.40/#section/Authentication), but the [code](https://github.com/docker/cli/blob/master/cli/command/registry.go#L50) is written such that the `X-Registry-Auth` header is expected to be URL safe (RFC 4648) base64.

* Agent reported state as "406" when modules are in backoff. (#3244)

To reproduce

1.    create a deployment which has a module which fails to launch, like an incorrect image name, or some weird mount Docker doesn't like.
2.    Look at the reported status of the edge runtime - If you are lucky, you will see the real error.
3.    Let the runtime run for long enough that the "OrderedRetryPlanRunner" puts the module on backoff.
4.    Look at the reported status of the edge runtime - You won't see the real error message from the failure - you will see a "406 -- The device is offline or not sending status reports"

The root cause is that every reconcile starts with status=Unknown(406), and when a module is in backoff, the status doesn't update.

So it goes:
reconcile loop 1: module command fails, error is reported
reconcile loop 2: module is in backoff now, Unknown is reported
...
reconcile loop n: module command is attempted again, error is reported
reconcile loop n+1: module is in backoff again, Unknown is reported.

User is more likely to see "406 -- The device is offline or not sending status reports" which is not cool.
This has been seen before, one example: https://github.com/Azure/iotedge/issues/2066

The fix here is to keep track of status in Agent and report current status until it changes (either via a new error

* Update dockerfiles for arm32 & amd64, update Helm charts.

* Fix teh edge agent tests.

Co-authored-by: Anca Antochi <ancan@microsoft.com>