Ordered start for authentication and mqtt bridge (#3307)
When edgeHub starts up, protocol heads start up parallel. In case of Authentication and mqtt bridge it could happen (and usually happened) that authentication came alive sooner. When a client connected after authentication came alive, but mqtt bridge had not, then the mqtt broker allowed the client to connect. When this client sent a message immediately, mqtt bridge was not subscribed to the necessary topics to handle the messages toward iothub. These messages got lost.
This change starts mqtt bride earlier and starts authentication only when everything is up. It does the reverse order on shutdown.