[PnP] ModuleClient test (#3316)
Adding PnP ModuleClient test to E2e tests. This test does a very similar thing that the deviceClient test does, but with a module.

It starts edge runtime and deploys LoadGen module. It gives LoadGen a modelId. 
LoadGen then sends some messages.
Once the test recognizes that messages have been received, it gets the modelId from the twin (using an iot hub endpoint - there is no service api for it at the moment) to validate that it is correct. 

Also had to modify loadGen slightly.
I also modified ModuleUtil and all the modules that use it. ModuleUtil should give the user the option to use ClientOptions, as it does with this PR.