[mqtt] Refactor topic translation and remove mqtt-broker-core crate (#3314)
We have a redundant `mqtt-broker-core` crate in the workspace. I was introduced by mistake so that `mqtt-broker` can reference `mqtt-edgehub` which is wrong dependency. It should be removed.
Now:
```
[mqtt-edgehub]  --> [mqtt-broker-core] <-- [mqtt-generic]
        ^              ^ 
        |              |
[mqtt-broker] ---------+
```
After change
```
[mqtt-edgehub] --> [mqtt-broker] <-- [mqtt-generic]
```

To do so I refactored incoming and outgoing tasks in connection and created `PacketProcessor` component. `PacketProcessor` converts from mqtt packets into broker messages back and forth. It allows to wrap packet processor in another one and deleted topic translation to this wrapper/adapter.