[PnP] Fixing bug that broke integ tests (#3365)
The last pnp PR broke our integ tests because of two subtle bugs.

First is that the metadataStore needs to be mocked in two more places.
2nd is that I had a more stringent check in MetadataStore that broke some of the integ tests, since the integ tests use empty string as edgeProductInfo.

As I don't want to cause any breaking changes, I changed the metadataStore check to be the same as it was before (CheckNotNull instead of CheckNonWhitespace)