Adding modelId to IoTHub device authentication flow (#3323)
Cherry pick from 3f891e8d115b1c48509a76564946ef35af47ce48
When a device does not have edgeHub set as the parent, edgeHub tries to connect to IoTHub for the device as a means of authentication. This allows the user to still connect a device through edgeHub even if the user does not want to (or just forgets) set the parent of the device as edgeHub.
We don't want to set the modelId of the auth until AFTER we know the device has been authenticated. Otherwise, someone could connect with a malicious or incorrect device, and edgeHub would save the modelId as the new modelId prematurely. Currently, if a user doesn't set edgeHub as the parent of a device, the modelId won't be stored for a device.
This PR passes the modelId to authentication through IoT Hub so that we can connect with the modelId when authenticating with IoT Hub.