Fix Subscription Processing Workaround (#3390) (#3409)
Removing a workaround that we added for 1.0.9 that fixed an issue we were having that when an MQTT identity with only subscriptions (no telemetry) reauthenticated, the identity wouldn't reconnect to the cloud. (PR of previous problem: https://github.com/Azure/iotedge/pull/3201)

The new way of doing this is to listen to when a device gets connected to EdgeHub and process subscriptions when it does.
To prevent a re-triggering of subscription processing when EdgeHub connects to the cloud on behalf of the device (this situation occurs during the device connecting to edgeHub flow), we use a ConcurrentDictionary to keep track of if subscription processing has already been kicked off for a DeviceConnecting. We check this in the other methods. 
This effectively ignores the Cloud Connection event trigger.

Cherry-picked from https://github.com/Azure/iotedge/pull/3390