Add option to bypass Edge installation for E2E tests (Linux) (#3421)
This change is needed for the support of Azure EFLoW (Edge For Linux on Windows).
The Mariner (Linux) VM that composes EFLoW comes with IoT Edge pre-provisioned.
This change allows E2E tests to omit installation of the _iotedge_ daemon and to use daemon pre-provisioned in the Mariner image.
This is an equivalent change for Linux to [#1188](https://github.com/Azure/iotedge/pull/1188).