Api proxy (#3490)
First version of the API proxy.
Feature wise it is complete:

able to receive a configuration from the twin and save it
monitor nginx and restart it
get certificates from workload API.
The generated image is ~22Mb.
Please see the readme on my account for more details:
https://github.com/huguesBouvier/iotedge/tree/iiot/edge-modules/api-proxy-module

What is left to do:

Completing the template for each use case. I still need to complete the template for pulling docker container image. But that is separate from the API proxy feature.
The readme is incomplete as I have not finished the template yet.