Refactor Command handler and add handle_authorized_identities event (#3474)
Refactoring CommandHandler to take multiple commands. Also, adding a new Command, handle_authorized_identities, that will intake authorized identities as a payload from edgeHub, convert it to ServiceIdentity objects and send that to the broker as a SystemEvent.

Currently, the broker just has a dummy implementation for this SystemEvent.

We do this because edgeHub does the authorization for all identities, and the broker just receives it from edgeHub whenever the edgeHub cache is updated.