[mqtt] Mitigate bad startup order (#3523)
I made a mistake overlooking that hot tasks do not work the C# way. This leads us to a state where the broker hangs and doesn't open its ports. The tests don't catch this because our test-util startup is not implemented the same way as our actual startup. 

This PR serves as a mitigation of the issue. I have a bug open on me to fix the underlying issue which will involve changing the startup logic.