Add `always_reprovision_on_startup` setting to DPS  provisioning configuration (#3435)
Before this change, when using any of the DPS provisioning methods,
iotedged would always reprovision with DPS on startup. It would maintain
a provisioning backup, but this backup was only used if the provisioning
with DPS failed for any reason.

If the new setting in this change is set to false, iotedged will first attempt
to restore the backup, and only reprovision with DPS if that fails.

The primary reason this was done was due to a quirk of DPS-TPM provisioning.
`iotedged`, or rather `utpm` as used by `libiothsm-std`, reserves a single slot
on the TPM to store keys. However DPS-TPM provisioning involves two keys that
must be loaded in the TPM - the initial challenge key and the final
device identity key. Thus the DPS-TPM provisioning method must store
the challenge key in the same slot that occupied
the previous device identity key. However, if the second part of
DPS-TPM provisioning fails and `iotedged` never receives
the new device identity key, the previous device identity key is
irrevocably lost even if the rest of the information can be restored from
the provisioning backup.

This creates a problem when modules request their SAS tokens to be signed
so they can talk to Edge Hub, `iotedged` will derive their module identity keys
using the device identity key that is actually the challenge key.
However Edge Hub will validate the SAS tokens against the module identity keys
that it received from IoT Hub, which were derived using the previous
device identity key. So Edge Hub will reject the modules' SAS tokens.

With this change, and if the setting has been set to false, `iotedged` won't
attempt the DPS-TPM provisioning at all, so the "previous" device identity key
in the TPM will remain.

One last thing to be handled carefully is the reprovisioning trigger.
If the device identity was deleted from IoT Hub, then before this change
the reprovision on startup would've caught it. With this change, the only way
it can be detected is when Edge Agent fails to connect to IoT Hub and thus
triggers a reprovision using the management API. However triggering the
reprovision using the management API is itself opt-in using
the `dynamic_reprovisioning` config flag. Since it doesn't make sense to have
a device that doesn't reprovision with the cloud but also doesn't reprovision
itself if it needs to, this change makes it so that setting
`always_reprovision_on_startup` to false implicitly sets
`dynamic_reprovisioning` to true.

Note that the `always_reprovision_on_startup` seting defaults to true
to retain the original behavior of always reprovisioning on startup,
for backward-compatiblity. However it is not expected that there would be
any problems from setting it to false.

Also note that this change makes `BackupProvisioning` generic over
the underlying provisioner (as it was always intended to be) rather than being
an internal detail of the three DPS provisioning methods. This means it could
be used for other provisioning methods like external in the future.
In this case, the `always_reprovision_on_startup` config would be moved
from being DPS-specific to applying to all provisioning methods. Luckily,
this can be done without any backward-incompatible changes to the config,
because the provisioning method parameters are already flattened, so that
the `always_reprovision_on_startup` flag is already at the top-level of
the provisioning settings.