[mqtt] Refactor command handler (#3542)
The code related to a command handler used to be in 2 separate modules. There are some changes to consolidate code for command in the one parent module. Each command handler is in its own module and `Command` trait is changed a bit to support custom Error type per command.
Removed unnecessary allocations (usually with strings).
`CommandHandler` is initialized in the bootstrap code.