Report provisioning type in metrics (#3543)
Iotedged's management API exposes `/systeminfo`, which returns information about the host, such as iotedged version, OS, architecture, and # CPUs. We also want it to return information about how the device is provisioned, to make it available on Edge agent's Prometheus endpoint (as part of the`host_information` tag in the `edgeAgent_metadata` gauge). This change adds `provisioning` to the `/systeminfo` response, and plumbs it through to the metric. It's an object that looks like:

```
{
  "type": string,
  "dynamicReprovisioning": boolean,  // default: false
  "alwaysReprovisionOnStartup": boolean  // optional, default: true
}
```

The `type` property can have one of the following values:

```
manual.device_connection_string
manual.x509
dps.tpm
dps.symmetric_key
dps.x509
external
```

The `alwaysReprovisionOnStartup` property is optional because it's (currently) only used for DPS provisioning.