Fix for wsl2 (#3574)
Built-in Moby engine is not used in LCOW configuration. However, the service remains active and consumes the DNS port (53) that caused WSL2 to fail (https://github.com/microsoft/WSL/issues/4929).

This PR makes two simple changes when running w/ LCOW:  (a) disables the built-in Moby engine  and (b) removes the iotedge service dependency. It also doesn't enable the "Containers" Windows feature which is only required for Windows containers.