Aggregate metrics before upload (#3482) (#3589)
This will aggregate metrics before the are uploaded. This is because many metrics have a high cardnality, since they have a new entry for each module and/or each rout input/output. 

since we hash those values anyway, pre-aggregating can save a significant amount of bandwidth.

It also updates netstandard to 2.1 to use Hashcode.Combine
Microsoft.Azure.WebJobs.Extensions.EdgeHub.csproj is not updated because it relies on older libraries.