[mqtt] RocksDB Persistence (#3549)
Adding support for rocksDB persistence. We previously only had in-memory persistence.

Previously, the plan was to have a persistor interface for in-memory and disk persistence implementations. I refactored the module structure to move the interface from the persistor to the waking_state. This way, we can better re-use code in that we only need one persistor/loader implementation.