[1.0.10] Set up edgeAgent container to run agent as non-root. (#3572)
This allows Agent to run as non-root in docker, and as `ContainerUser` in Windows.

## WIndows

We have set the default user in the Windows image to be `ContainerUser`.

## Linux
We have a default user (edgeagentuser, default Linux UID 1000) available if the user wants to start the container as non-root.  This will be created when the edgeAgent image is built.  However, if the user does not want to use UID 1000, they can start the image as root, set `EDGEAGENTUSER_ID` to the value that they want, and the container's script will create the user in the container and setup the correct permissions for this user to run edgeAgent. 

Some notes:

### Linux UID and certificate stores
Edge agent wants to install a trust bundle into a store. The dotnet core code gives us 2 location options for `X509Store`, `LocalMachine` or `CurrentUser`. The former needs write access to system directories and the latter needs a known user's home directory. The agent code uses the `CurrentUser` location. In order for the Agent to run as non-root, the current user's UID needs to have a writable home directory associated with it.  I chose to create UID 1000 when doing this for Edge on K8s, and I chose to continue to use this as the default for running as non-root on docker.

Regardless of the UID chosen for the container, this **must match** the UID ownership of the management socket file to allow edgeAgent access to the management socket.  If the user runs the agent as non-root through Docker settings, the management socket must be set.

### Setting UID on management socket.

Example using UID 1000.

**Ubuntu/debian 9 systems**
```bash
sudo systemctl edit iotedge.mgmt.socket
# add the following lines
```
> [Socket]
> SocketUser=1000
```bash
#save the override file.
sudo systemctl daemon-reload
sudo systemctl restart iotedge.mgmt.socket
```

**On other systems**
```bash
sudo chown 1000:iotedge /var/lib/iotedge/mgmt.sock
```