Revert "[mqtt] RocksDB Persistence (#3549)" (#3631)
This reverts commit b8800a2b98916fd779b234ae70b8674f3c2c1e2f.

The rocksdb build does not work right now on musl (needed for alpine). There is some conflicting information about if it is possible to get it building on alpine, but there is some conflicting info. I have overcame two issues people have ran into, but there is a linking problem that I can't yet figure out.

Here are some links for context:
- People say it can work: https://github.com/rust-rocksdb/rust-rocksdb/issues/174#issuecomment-409410115
- Someone with same linking problem says it won't work: https://github.com/rust-rocksdb/rust-rocksdb/issues/418

I am going to do some more investigation then sync with Arnav and David to see if fixable.