Using self-host Windows Agent for E2E test framework (#3625)
When the end-to-end tests run on Windows hosted agents, all tests with modules that connect to edge hub fail. Tests that are only concerned with provisioning iotedged, or that use leaf devices to talk to edge hub, pass. The module-based tests all fail with:
```
Microsoft.Azure.Devices.Client.Exceptions.IotHubCommunicationException: Transient network error occurred, please retry.
 ---> System.Net.Sockets.SocketException (11001): No such host is known.
```
It has been difficult to figure out what's happening on the hosted agents since we can't debug them. We know the problem doesn't happen on custom Windows agents, so this change switches the pipeline to custom agents.