Add metric to verify if edge device is running in some virtualized environment. (#3624)
A new metric - Virtualized - was added to indicate if the edge device is running in some virtualized environment. We're not collecting details about the virtualized environment like product name.
This metric will be collect only for Linux machines using the command systemd-detect-virt (https://www.freedesktop.org/software/systemd/man/systemd-detect-virt.html).
The new metric was added in the model SystemInfo.
To verify the new field check the edgeagent_metadata metric.