Add connectivity handler (#3699)
Adding handler for connectivity events. 
The idea is to create the channel in the bridge and the sender is passed to upstream pump and receiver to downstream pump. This will be in a different PR since changes are based on the pump.

Note: PR has to be merged after changes to mqtt3 client changes.