Allow test context without container registry info (#3714)
The service deployment pipeline (`builds/service/service-deployment.yaml`) runs a single end-to-end test ("TempSensor") in an Azure pipeline. The IoT Hub team uses it to validate new deployments of their service. Since the pipeline only runs a single, basic test, it's context.json file is very simple:
```json
{
  "logFile": "/path/to/testoutput.log"
}
```
The pipeline has been running against 1.0.9, but now that 1.0.10 is released I tried to upgrade it. Something has changed during this release, and the "registries" object in context.json is no longer optional, causing a thrown exception in the fixtures. This change restores the correct behavior.