[mqtt] Outgoing and incoming message pump for bridge (#3700)
Currently the bridge only supports connecting to a broker and ingesting messages into a store. We need to support outgoing and incoming message pumps to actually send messages.

This PR does that by refactoring the BridgeController and Bridge. It also adds a Pump component that is responsible for connecting to a broker and setting up the message ingress/egress.