Fix metrics upload (#3737)
One of the changes to edgeAgent cause the SendEventBatchAsync function of the module sdk to stop working correctly. This pr removes the call to SendEventBatchAsync and replaces it with multiple calls to SendEventAsync. 

The main difference between the 2 is that SendEventBatchAsync is transactional, while iterative SendEventAsync is not. Because the metrics worker will do a full retry on upload failure, this means that if the metrics upload fails partway through a batch, metrics could be duplicated. I am not worried about this in practice, because:
* Metrics being batched should be fairly rare, only happening when a device is offline for an extended period of time (>3 days)
* In addition to batching being rare, the upload still needs to fail partway through a batch for data to be duplicated.
* Any duplicate data is easy to de-dupe since it has a unique instance guid and timestamp