Upstream proxy for broker (vs c# SDK Client/ModuleClient) (#3612)
This is a draft PR to show how the upstream communication is going to work through the MQTT broker. There are two main points:

1) How to switch from 'legacy upstream' to 'brokered upstream'

The mqtt adapter already has a setting group (mqttBrokerSettings in the appsettings file). This got a new flag called "legacyUpstream" (which is true by default, ensuring that everything works as before). Setting it false will configure the dependency injection to create the proxy classes that routes the communication through the broker

2) How the proxy classes are implemented

There are 4 important classes, 2 of them is just kept because the existing infrastructure builds on them. These classes are BrokeredClouldConnection/BrokeredCloudConnectionProvider/BrokeredCloudProxy and BrokeredClouldProxyDispatcher

The most important class is the BrokeredCloudProxyDispatcher, which is a singleton service. This class is plugged in the MqttBrokerAdapter infrastructure by implementing IMessageConsumer/IMessageProducer. Implementing those interfaces the mqtt connector class (which does the mqtt "talk") sends incoming messages to BrokeredCloudProxyDispatcher.

BrokeredCloudProxy, which represents an upstream proxy for the client to edgeHub forwards every call to BrokeredCloudProxyDispatcher, adding its identity to the call. This way BrokeredCloudProxyDispatcher handles all operations for all connected clients.