Fix Edge identity creation in X509 thumbprint E2E test (#3775)
The secondary thumbprint is mandatory when creating an X509 self-signed identity, but it seems that the iothub API implementation doesn't enforce it.  So the testcase ended up creating a non-compliant Edge device by not having both thumprints.  This causes initial cache refresh to fail when we parse the ScopeResults from iothub, and prevent the downstream device from connecting.