[k8s] Environment Variables can be null or empty. (#3780)
* Environment Variables can be null or empty.
For JSON, use a deserialization function to read env vars as optional,
then transform empty entries into blank values.
For YAML config, don't return an error in Yaml::Null case,
transform it into ValueKind::Nil

* Just return ValueKind::Nil for all "bad" cases.

Fix derserialize.
We want to allow nulls for env variables.
This is treated as the equivalent of an empty string in the docker spec,
because both are interpreted as such,
`["IS_AN_EMPTY_STRING=", "IS_A_NULLSTRING="]`
In order to avoid copying keys and values, I am deserializing as a HashMap
and then using `drain()` to transform and collect it into a BTreeMap.