Resync leaf identities if an Edge gets re-parented (#3788)
When adding a new device that doesn't have its identity sync'ed down yet, we rely on the bottom-most Edge to send a refresh request, which would trigger a corresponding refresh on every Edge up the chain until the root sends the request to IoTHub.  Then on the way back every Edge would have the new device identity added to the cache.

But in case of reparenting, the bottom-most Edge will already have the leaf device identity, whereas the new parent Edge will not.  So the authentication will fail unless the parent Edge has logic to refresh the leaf device identity OnBehalfOf the bottom-most Edge.